feat(build): 添加 SHA-NI 指令集支持和代码警告修复
- CMakeLists.txt: 添加 -msha 编译选项启用 SHA-NI 指令集
- sha256.cpp: 修复 old-style cast, 添加 SHA-NI 代码路径 (待验证)
- sha512.cpp: 修复 old-style cast 和 sign-conversion 警告
- sha3.cpp: 修复 sign-conversion 警告，使用 size_t 循环变量
- chacha20_poly1305.cpp: 修复 old-style cast 和 sign-conversion 警告
- mac.cpp: 修复 conversion 警告，使用 static_cast
- sm3.cpp: 修复 old-style cast 警告
- sm4.cpp: 修复 sign-conversion 警告
- zuc.cpp: 修复 sign-conversion 和 conversion 警告
- aes.cpp: 修复 old-style cast 警告

性能基准测试结果 (vs OpenSSL 3.3.1):
- BLAKE2b: +30% FASTER (992 MB/s vs 761 MB/s)
- BLAKE2s: +39% FASTER (632 MB/s vs 455 MB/s)
- SHA-256: 16% (scalar 实现, SHA-NI 待修复)
- SHA3-256: 79%
- SM3: 78%

+++ src/crypto/aes.cpp
@@ -293,17 +293,17 @@ static const uint32_t TTABLE_RCON[15] = {
 static inline uint8_t gf_mul(uint8_t a, uint8_t b) {
     uint8_t result = 0;
     uint8_t temp = a;
-    for (int i = 0; i < 8; i++) {
-        uint8_t mask = (uint8_t)(-(int8_t)((b >> i) & 1));
+    for (size_t i = 0; i < 8; i++) {
+        uint8_t mask = static_cast<uint8_t>(-static_cast<int8_t>((b >> i) & 1));
         result ^= (temp & mask);
-        uint8_t hi_bit = (uint8_t)((temp >> 7) & 1);
-        temp = (uint8_t)((temp << 1) ^ (0x1b & (uint8_t)(-((int8_t)hi_bit))));
+        uint8_t hi_bit = static_cast<uint8_t>((temp >> 7) & 1);
+        temp = static_cast<uint8_t>((temp << 1) ^ (0x1b & static_cast<uint8_t>(-static_cast<int8_t>(hi_bit))));
     }
     return result;
 }
 
 static inline void xor_block(uint8_t* out, const uint8_t* a, const uint8_t* b) {
-    for (int i = 0; i < 16; i++) {
+    for (size_t i = 0; i < 16; i++) {
         out[i] = a[i] ^ b[i];
     }
 }
